pyproject.toml
[tool.poetry]
name = "plotting-server"
version = "0.1.0"
description = ""
authors = ["Antonello Pellecchia <antonello.pellecchia@cern.ch>"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
matplotlib = "^3.4.3"
PyVISA = "^1.11.3"
seaborn = "^0.11.2"
PyYAML = "^6.0"
bottle = "^0.12.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
